requirements.txt
streamlit>=1.31.0
langchain-core<0.3
langchain-community<0.3
langchain-openai==0.1.7
langchain>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
bs4>=0.0.2
html5lib>=1.1
lxml>=4.9.0
